requirements.txt
transformers
datasets
torch
python-docx
scikit-learn
tqdm
